requirements.txt
fastapi==0.62.0
uvicorn==0.13.1
pymongo
python-decouple